+++ lib/lindi.dart
@@ -1,6 +1,7 @@
 library;
 
 export 'src/core/lindi_view_model.dart';
+export 'src/core/lindi_storage_view_model.dart';
 export 'src/core/lindi.dart';
 export 'src/builders/lindi_builder.dart';
 export 'src/utils/lindi_exception.dart';